Controller tests

+++ src/SFA.DAS.Approvals.Api/Controllers/CohortController.cs
@@ -3,8 +3,6 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using SFA.DAS.Approvals.Application.Cohorts.Queries;
 using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 
 namespace SFA.DAS.Approvals.Api.Controllers
@@ -25,7 +23,7 @@ namespace SFA.DAS.Approvals.Api.Controllers
 
         [HttpGet]
         [Route("{cohortId}")]
-        public async Task<IActionResult> Get(int cohortId)
+        public async Task<IActionResult> Get(long cohortId)
         {
             try
             {
